三層アーキテクチャ構造の取り込み
