--- .github/PULL_REQUEST_TEMPLATE.md
@@ -1,34 +1,0 @@
-## 変更内容の説明
-
-<!-- このプルリクエストで何を変更したかを簡潔に説明してください -->
-
-## 変更の種類
-
-- [ ] バグ修正
-- [ ] 新機能
-- [ ] 破壊的変更
-- [ ] ドキュメント更新
-- [ ] リファクタリング
-- [ ] テストの追加・修正
-
-## テスト
-
-- [ ] 新しいテストを追加した
-- [ ] 既存のテストが通ることを確認した
-- [ ] 手動テストを実行した
-
-## チェックリスト
-
-- [ ] コードは自己文書化されている、またはコメントが適切に追加されている
-- [ ] 関連するドキュメントを更新した
-- [ ] CI/CDパイプラインが通ることを確認した
-- [ ] セキュリティの考慮事項を検討した
-
-## 関連Issue
-
-<!-- 関連するIssueがある場合はここに記載 -->
-Closes #
-
-## その他
-
-<!-- その他に伝えたいことがあれば記載 -->
