--- .github/workflows/ci.yml
@@ -1,84 +1,0 @@
-name: CI/CD Pipeline
-
-on:
-  push:
-    branches: [ main, develop, draque, CICD_test ]
-  pull_request:
-    branches: [ main, develop, draque, CICD_test ]
-
-permissions:
-  contents: read
-  checks: write
-  pull-requests: write
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    name: Test and Build
-    
-    steps:
-    - name: Checkout code
-      uses: actions/checkout@v4
-      
-    - name: Set up JDK 17
-      uses: actions/setup-java@v4
-      with:
-        java-version: '17'
-        distribution: 'temurin'
-        
-    - name: Cache Maven dependencies
-      uses: actions/cache@v4
-      with:
-        path: ~/.m2
-        key: ${{ runner.os }}-m2-${{ hashFiles('**/pom.xml') }}
-        restore-keys: ${{ runner.os }}-m2
-        
-    - name: Run tests
-      run: mvn clean test
-      
-    - name: Generate test report
-      uses: dorny/test-reporter@v1
-      if: always() && (github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name == github.repository)
-      with:
-        name: Maven Tests
-        path: target/surefire-reports/*.xml
-        reporter: java-junit
-        
-    - name: Build application
-      run: mvn clean compile
-      
-    - name: Package application
-      run: mvn clean package -DskipTests
-      
-    - name: Upload build artifacts
-      uses: actions/upload-artifact@v4
-      with:
-        name: jar-artifact
-        path: target/*.jar
-
-  docker-build:
-    runs-on: ubuntu-latest
-    needs: test
-    name: Docker Build Test
-    
-    steps:
-    - name: Checkout code
-      uses: actions/checkout@v4
-      
-    - name: Set up Docker Buildx
-      uses: docker/setup-buildx-action@v3
-      
-    - name: Build Docker image
-      uses: docker/build-push-action@v5
-      with:
-        context: .
-        push: false
-        tags: draque:test
-        cache-from: type=gha
-        cache-to: type=gha,mode=max
-        
-    - name: Test Docker Compose
-      run: |
-        echo "GOOGLE_MAPS_API_KEY=test-key" > .env
-        docker compose config
-        docker compose build
