--- .github/workflows/code-quality.yml
@@ -1,50 +1,0 @@
-name: Code Quality Analysis
-
-on:
-  push:
-    branches: [ main, develop, draque, CICD_test ]
-  pull_request:
-    branches: [ main, develop, draque, CICD_test ]
-
-jobs:
-  code-quality:
-    runs-on: ubuntu-latest
-    name: Code Quality & Coverage
-    
-    steps:
-    - name: Checkout code
-      uses: actions/checkout@v4
-      with:
-        fetch-depth: 0  # Shallow clones should be disabled for better SonarCloud analysis
-      
-    - name: Set up JDK 17
-      uses: actions/setup-java@v4
-      with:
-        java-version: '17'
-        distribution: 'temurin'
-        
-    - name: Cache Maven dependencies
-      uses: actions/cache@v4
-      with:
-        path: ~/.m2
-        key: ${{ runner.os }}-m2-${{ hashFiles('**/pom.xml') }}
-        restore-keys: ${{ runner.os }}-m2
-        
-    - name: Run tests with coverage
-      run: mvn clean test jacoco:report
-      
-    - name: Generate coverage reports for Codecov
-      uses: codecov/codecov-action@v3
-      with:
-        file: ./target/site/jacoco/jacoco.xml
-        fail_ci_if_error: false
-        
-    - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v3
-      if: always()
-      with:
-        token: ${{ secrets.CODECOV_TOKEN }}
-        file: ./target/site/jacoco/jacoco.xml
-        flags: unittests
-        name: codecov-umbrella
-        fail_ci_if_error: false
