--- .github/workflows/railway-deploy.yml
@@ -1,22 +1,0 @@
-name: Deploy to Railway
-
-on:
-  push:
-    branches: [ main, develop ]
-
-jobs:
-  deploy:
-    runs-on: ubuntu-latest
-    if: github.event_name == 'push' && (github.ref == 'refs/heads/main' || github.ref == 'refs/heads/develop')
-    
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-
-      - name: Deploy to Railway
-        run: |
-          curl -fsSL https://railway.app/install.sh | sh
-          railway login --token ${{ secrets.RAILWAY_TOKEN }}
-          railway up --service ${{ secrets.RAILWAY_SERVICE_NAME }}
-        env:
-          RAILWAY_TOKEN: ${{ secrets.RAILWAY_TOKEN }}
