--- railway.json
@@ -1,17 +1,0 @@
-{
-  "$schema": "https://railway.app/railway.schema.json",
-  "build": {
-    "builder": "DOCKERFILE"
-  },
-  "deploy": {
-    "numReplicas": 1,
-    "restartPolicyType": "ON_FAILURE",
-    "restartPolicyMaxRetries": 10
-  },
-  "volumes": [
-    {
-      "name": "config",
-      "mountPath": "/config"
-    }
-  ]
-}
