+++ src/main/java/com/example/dockerapi/dto/EncounterResponse.java
@@ -1,0 +1,30 @@
+package com.example.dockerapi.dto;
+
+public class EncounterResponse {
+    private boolean appear;
+    private MonsterDto monster;
+
+    public EncounterResponse() {
+    }
+
+    public EncounterResponse(boolean appear, MonsterDto monster) {
+        this.appear = appear;
+        this.monster = monster;
+    }
+
+    public boolean isAppear() {
+        return appear;
+    }
+
+    public void setAppear(boolean appear) {
+        this.appear = appear;
+    }
+
+    public MonsterDto getMonster() {
+        return monster;
+    }
+
+    public void setMonster(MonsterDto monster) {
+        this.monster = monster;
+    }
+}
